requirements.txt
PyQt5==5.15.9
torch==2.0.0
torchvision==0.15.0
timm==0.9.2
Pillow==9.5.0
numpy==1.24.3
